requirements.txt
# Minimal requirements for MonitCam
pyautogui
opencv-python
numpy
pillow
pywin32
mss
flask
